# Fix queue contention